+++ collect_app/src/main/java/org/odk/collect/android/injection/config/AppDependencyModule.java
@@ -53,9 +53,9 @@ import org.odk.collect.android.formmanagement.FormsDataService;
 import org.odk.collect.android.formmanagement.OpenRosaClientProvider;
 import org.odk.collect.android.geo.MapConfiguratorProvider;
 import org.odk.collect.android.geo.MapFragmentFactoryImpl;
-import org.odk.collect.android.instancemanagement.InstanceSubmitter;
+import org.odk.collect.android.instancemanagement.send.InstanceSubmitter;
 import org.odk.collect.android.instancemanagement.InstancesDataService;
-import org.odk.collect.android.instancemanagement.autosend.AutoSendSettingsProvider;
+import org.odk.collect.android.instancemanagement.send.autosend.AutoSendSettingsProvider;
 import org.odk.collect.android.instancemanagement.send.InstanceUploader;
 import org.odk.collect.android.instancemanagement.send.OpenRosaServerInstanceUploader;
 import org.odk.collect.android.instancemanagement.send.ReadyToSendViewModel;
