+++ collect_app/src/main/java/org/odk/collect/android/instancemanagement/send/InstanceUploaderActivity.java
@@ -29,9 +29,7 @@ import androidx.lifecycle.ViewModelProvider;
 import org.odk.collect.android.activities.FormFillingActivity;
 import org.odk.collect.android.fragments.dialogs.SimpleDialog;
 import org.odk.collect.android.injection.DaggerUtils;
-import org.odk.collect.android.instancemanagement.InstanceUploadResultKt;
 import org.odk.collect.android.instancemanagement.InstancesDataService;
-import org.odk.collect.android.instancemanagement.InstanceUploadResult;
 import org.odk.collect.android.projects.ProjectsDataService;
 import org.odk.collect.android.utilities.ApplicationConstants;
 import org.odk.collect.android.utilities.ArrayUtils;
