+++ collect_app/src/main/java/org/odk/collect/android/instancemanagement/send/autosend/InstanceAutoSendFetcher.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.instancemanagement.autosend
+package org.odk.collect.android.instancemanagement.send.autosend
 
 import org.odk.collect.forms.Form
 import org.odk.collect.forms.FormsRepository
