+++ collect_app/src/test/java/org/odk/collect/android/instancemanagement/InstancesDataServiceTest.kt
@@ -10,6 +10,7 @@ import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.verifyNoInteractions
 import org.mockito.kotlin.whenever
+import org.odk.collect.android.instancemanagement.send.InstanceSubmitter
 import org.odk.collect.android.notifications.Notifier
 import org.odk.collect.android.projects.ProjectDependencyModule
 import org.odk.collect.android.utilities.ChangeLocks
@@ -60,7 +61,8 @@ class InstancesDataServiceTest {
     private val httpInterface = mock<OpenRosaHttpInterface>()
     private val notifier = mock<Notifier>()
 
-    private val instanceSubmitter = InstanceSubmitter(projectsDependencyModuleFactory, mock(), httpInterface)
+    private val instanceSubmitter =
+        InstanceSubmitter(projectsDependencyModuleFactory, mock(), httpInterface)
 
     private val instancesDataService =
         InstancesDataService(
