+++ collect_app/src/test/java/org/odk/collect/android/utilities/FormsUploadResultInterpreterTest.kt
@@ -7,7 +7,7 @@ import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.`is`
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.odk.collect.android.instancemanagement.InstanceUploadResult
+import org.odk.collect.android.instancemanagement.send.InstanceUploadResult
 import org.odk.collect.android.instancemanagement.send.FormUploadException
 import org.odk.collect.forms.instances.Instance
